Optimized sructure of the project
