+++ controllers/index.js
@@ -1,0 +1,3 @@
+const contactsControllers = require("./contacts-controllers");
+
+module.exports = { contactsControllers };
