+++ models/contacts.json
@@ -58,11 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "Wiyz2cCYdObsD0ncGI3GO",
-    "name": "Alex",
-    "email": "abc@abc.abc",
-    "phone": "+1234567890"
   }
 ]
