+++ schemas/index.js
@@ -1,0 +1,3 @@
+const { contactAddSchema } = require("./joi-schemas");
+
+module.exports = { contactAddSchema };
